+++ src/utils/entities.js
@@ -4,7 +4,7 @@ const {
   STATUSES: { UNKNOWN, CRITICAL, WARNING, SUCCESS },
 } = StatusIcon;
 
-export const entityStatus = ({ alertSeverity }) => {
+export const entityStatus = ({ alertSeverity } = {}) => {
   switch (alertSeverity) {
     case 'NOT_ALERTING': {
       return SUCCESS;
@@ -20,3 +20,27 @@ export const entityStatus = ({ alertSeverity }) => {
     }
   }
 };
+
+export const guidsToArray = (guids = {}, maxArrayLen = 10) =>
+  Object.keys(guids).reduce((acc, type) => {
+    const typeGuids = guids[type];
+    if (!typeGuids || !typeGuids.length) return acc;
+    return [
+      ...acc,
+      ...Array.from(
+        { length: Math.ceil(typeGuids.length / maxArrayLen) },
+        (_, i) => {
+          const startIdx = i * maxArrayLen;
+          return typeGuids.slice(startIdx, startIdx + maxArrayLen);
+        }
+      ),
+    ];
+  }, []);
+
+export const entitiesDetailsFromQueryResults = (res = {}) =>
+  Object.keys(res).reduce((acc, cur) => {
+    const signalsArray = res[cur];
+    if (!Array.isArray(signalsArray)) return acc;
+    signalsArray.forEach(({ guid, name }) => (acc[guid] = { name }));
+    return acc;
+  }, {});